unix: use uv_signal_*() functions
Makes it possible for multiple SIGUSR2 handlers to coexist peacefully.
Consider this snippet:

    var heapdump = require('heapdump');
    process.on('SIGUSR2', function() {
      // ...
    });

Before this commit, a snippet like the one below would stop heapdump
from working because the event listener overrides the signal handler
that heapdump installs.

This change makes node-heapdump incompatible with node.js <= 0.8 as it
uses a libuv API that is only available in node.js >= 0.10.

Fixes #13.