src: fix windows-only memory leak
Call v8::HeapSnapshot::Delete() to free the memory.  Doing so requires
a const_cast<> that is currently safe.  Let's hope it stays that way
with future V8 releases.

The memory leak only affects Windows.  On UNIX platforms, the snapshot
is created in a forked process that exits after writing it to disk.

Fixes #24.
Fixes #27.