lib,src: make nofork the default behavior
Fork-before-snapshotting is fundamentally incompatible with newer V8
releases.  V8 uses threads now but those don't survive the call to
fork().  Then, when creating the snapshot, V8 tries to sync up with
a thread that no longer exists and deadlocks.

Another issue is that forking breaks the comparison view in DevTools
because the heap object IDs don't match up.  V8 updates them after
creating the snapshot but those changes don't make it back to the
parent process.

Fixes #37 and #41.